refactor(describegpt): extract dictionary cluster into submodule
Move the neuro-procedural dictionary cluster — DictionaryEntry,
StatsRecord, FrequencyRecord, plus parse_stats_csv, parse_frequency_csv,
generate_code_based_dictionary, combine_dictionary_entries, and
parse_llm_dictionary_response — into src/cmd/describegpt/dictionary.rs.

These functions form a cohesive unit: they parse the stats/frequency
CSVs into typed records and then mechanically derive the dictionary
(with the LLM only filling in label + description). Their only coupling
to the rest of describegpt is parse_llm_dictionary_response's call to
extract_json_from_output, imported via super::.

Struct fields are pub(super) so sibling submodules can read them;
formatters.rs updated to import from super::dictionary::DictionaryEntry.

No behavior change. Main file drops from 5,576 to 5,062 lines.
Clippy clean; all 64 integration + 7 unit tests pass.
