+++ src/cmd/describegpt.rs
@@ -362,7 +362,7 @@ use cached::{
     DiskCache, IOCached, RedisCache, Return, proc_macro::io_cached, stores::DiskCacheBuilder,
 };
 use foldhash::{HashMap, HashMapExt, HashSet};
-use indexmap::{IndexMap, IndexSet};
+use indexmap::IndexSet;
 use minijinja::{Environment, context};
 use reqwest::blocking::Client;
 use serde::{Deserialize, Serialize};
@@ -382,7 +382,13 @@ use crate::{
 #[cfg(feature = "feature_capable")]
 use crate::{lookup, lookup::LookupTableOptions};
 
+mod dictionary;
 mod formatters;
+
+use dictionary::{
+    combine_dictionary_entries, generate_code_based_dictionary, parse_frequency_csv,
+    parse_llm_dictionary_response, parse_stats_csv,
+};
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, EnumString, Display)]
 #[strum(ascii_case_insensitive)]
@@ -600,43 +606,6 @@ struct PhaseResponse {
     response:    String,
     reasoning:   String,
     token_usage: TokenUsage,
-}
-
-// Data structures for neuro-procedural dictionary generation
-#[derive(Debug, Clone, Serialize, Deserialize)]
-struct DictionaryEntry {
-    name:        String,
-    r#type:      String,
-    label:       String,
-    description: String,
-    min:         String, // Empty string if not available
-    max:         String, // Empty string if not available
-    cardinality: u64,
-    enumeration: String, // Empty string if not enumerable, otherwise values on separate lines
-    null_count:  u64,
-    addl_cols:   IndexMap<String, String>, // Addl columns from stats (preserves order)
-    examples:    String,                   // Format: "val1 [cnt1], ... or "<ALL_UNIQUE>"
-}
-
-// Helper structs for parsing CSV data
-#[derive(Debug, Clone)]
-struct StatsRecord {
-    field:       String,
-    r#type:      String,
-    cardinality: u64,
-    nullcount:   u64,
-    min:         String,                   // Empty string if not available
-    max:         String,                   // Empty string if not available
-    addl_cols:   IndexMap<String, String>, // Addl columns from stats CSV (preserves order)
-}
-
-#[derive(Debug, Clone)]
-struct FrequencyRecord {
-    field:      String,
-    value:      String,
-    count:      u64,
-    percentage: f64,
-    rank:       f64,
 }
 
 // environment variables
@@ -1138,335 +1107,6 @@ fn get_prompt_file(args: &Args) -> CliResult<&PromptFile> {
     }
 }
 
-/// Parse stats CSV into structured records
-/// Returns the records and the ordered list of additional column names (in CSV order)
-fn parse_stats_csv(stats_csv: &str) -> CliResult<(Vec<StatsRecord>, Vec<String>)> {
-    let mut rdr = csv::ReaderBuilder::new()
-        .has_headers(true)
-        .from_reader(stats_csv.as_bytes());
-
-    let headers = rdr.headers()?.clone();
-
-    // Standard column names that we handle explicitly
-    let std_cols: HashSet<&str> = ["field", "type", "cardinality", "nullcount", "min", "max"]
-        .iter()
-        .copied()
-        .collect();
-
-    // Find column indices for standard columns
-    let field_idx = headers
-        .iter()
-        .position(|h| h == "field")
-        .ok_or_else(|| CliError::Other("Stats CSV missing 'field' column".to_string()))?;
-
-    let type_idx = headers
-        .iter()
-        .position(|h| h == "type")
-        .ok_or_else(|| CliError::Other("Stats CSV missing 'type' column".to_string()))?;
-
-    let cardinality_idx = headers.iter().position(|h| h == "cardinality");
-    let nullcount_idx = headers
-        .iter()
-        .position(|h| h == "nullcount")
-        .ok_or_else(|| CliError::Other("Stats CSV missing 'nullcount' column".to_string()))?;
-    let min_idx = headers.iter().position(|h| h == "min");
-    let max_idx = headers.iter().position(|h| h == "max");
-
-    // Collect indices of additional (non-standard) columns
-    let addl_col_indices: Vec<(usize, String)> = headers
-        .iter()
-        .enumerate()
-        .filter_map(|(idx, header)| {
-            if std_cols.contains(header) {
-                None
-            } else {
-                Some((idx, header.to_string()))
-            }
-        })
-        .collect();
-
-    let mut records = Vec::new();
-
-    for result in rdr.records() {
-        let record = result?;
-        let field = record
-            .get(field_idx)
-            .ok_or_else(|| CliError::Other("Stats CSV record missing field value".to_string()))?
-            .to_string();
-
-        let r#type = record
-            .get(type_idx)
-            .ok_or_else(|| CliError::Other("Stats CSV record missing type value".to_string()))?
-            .to_string();
-
-        let cardinality = cardinality_idx
-            .and_then(|idx| record.get(idx))
-            .and_then(|s| s.parse::<u64>().ok())
-            .unwrap_or(0);
-
-        let nullcount = record
-            .get(nullcount_idx)
-            .ok_or_else(|| CliError::Other("Stats CSV record missing nullcount value".to_string()))?
-            .parse::<u64>()
-            .map_err(|e| CliError::Other(format!("Failed to parse nullcount: {e}")))?;
-
-        let min = min_idx
-            .and_then(|idx| record.get(idx))
-            .map(std::string::ToString::to_string)
-            .unwrap_or_default();
-
-        let max = max_idx
-            .and_then(|idx| record.get(idx))
-            .map(std::string::ToString::to_string)
-            .unwrap_or_default();
-
-        // Collect additional columns, preserving CSV order
-        // Ensure all cols are present (w/ empty string if missing) to maintain consistent order
-        let mut addl_cols = IndexMap::new();
-        for (idx, col_name) in &addl_col_indices {
-            let value = record
-                .get(*idx)
-                .map(std::string::ToString::to_string)
-                .unwrap_or_default();
-            addl_cols.insert(col_name.clone(), value);
-        }
-
-        records.push(StatsRecord {
-            field,
-            r#type,
-            cardinality,
-            nullcount,
-            min,
-            max,
-            addl_cols,
-        });
-    }
-
-    // Extract ordered column names (preserving CSV order)
-    let ordered_col_names: Vec<String> =
-        addl_col_indices.into_iter().map(|(_, name)| name).collect();
-
-    Ok((records, ordered_col_names))
-}
-
-/// Parse frequency CSV into structured records
-fn parse_frequency_csv(frequency_csv: &str) -> CliResult<Vec<FrequencyRecord>> {
-    let mut rdr = csv::ReaderBuilder::new()
-        .has_headers(true)
-        .from_reader(frequency_csv.as_bytes());
-
-    let headers = rdr.headers()?.clone();
-
-    // Find column indices
-    let field_idx = headers
-        .iter()
-        .position(|h| h == "field")
-        .ok_or_else(|| CliError::Other("Frequency CSV missing 'field' column".to_string()))?;
-
-    let value_idx = headers
-        .iter()
-        .position(|h| h == "value")
-        .ok_or_else(|| CliError::Other("Frequency CSV missing 'value' column".to_string()))?;
-
-    let count_idx = headers
-        .iter()
-        .position(|h| h == "count")
-        .ok_or_else(|| CliError::Other("Frequency CSV missing 'count' column".to_string()))?;
-
-    let percentage_idx = headers
-        .iter()
-        .position(|h| h == "percentage")
-        .ok_or_else(|| CliError::Other("Frequency CSV missing 'percentage' column".to_string()))?;
-
-    let rank_idx = headers
-        .iter()
-        .position(|h| h == "rank")
-        .ok_or_else(|| CliError::Other("Frequency CSV missing 'rank' column".to_string()))?;
-
-    let mut records = Vec::new();
-
-    for result in rdr.records() {
-        let record = result?;
-        let field = record
-            .get(field_idx)
-            .ok_or_else(|| CliError::Other("Frequency CSV record missing field value".to_string()))?
-            .to_string();
-
-        let value = record
-            .get(value_idx)
-            .ok_or_else(|| CliError::Other("Frequency CSV record missing value".to_string()))?
-            .to_string();
-
-        let count = record
-            .get(count_idx)
-            .ok_or_else(|| CliError::Other("Frequency CSV record missing count".to_string()))
-            .and_then(|s| {
-                if s.is_empty() {
-                    Ok(0)
-                } else {
-                    s.parse::<u64>().map_err(|e| {
-                        CliError::Other(format!("Failed to parse count in frequency CSV: {e}"))
-                    })
-                }
-            })?;
-
-        let percentage = record
-            .get(percentage_idx)
-            .ok_or_else(|| CliError::Other("Frequency CSV record missing percentage".to_string()))
-            .and_then(|s| {
-                if s.is_empty() {
-                    Ok(0.0)
-                } else {
-                    s.parse::<f64>().map_err(|e| {
-                        CliError::Other(format!("Failed to parse percentage in frequency CSV: {e}"))
-                    })
-                }
-            })?;
-
-        let rank = record
-            .get(rank_idx)
-            .ok_or_else(|| CliError::Other("Frequency CSV record missing rank".to_string()))
-            .and_then(|s| {
-                if s.is_empty() {
-                    Ok(0.0)
-                } else {
-                    s.parse::<f64>().map_err(|e| {
-                        CliError::Other(format!("Failed to parse rank in frequency CSV: {e}"))
-                    })
-                }
-            })?;
-
-        records.push(FrequencyRecord {
-            field,
-            value,
-            count,
-            percentage,
-            rank,
-        });
-    }
-
-    Ok(records)
-}
-
-/// Generate code-based dictionary entries from stats and frequency data
-fn generate_code_based_dictionary(
-    stats_records: &[StatsRecord],
-    frequency_records: &[FrequencyRecord],
-    enum_threshold: usize,
-    num_examples: u16,
-    truncate_str: usize,
-    addl_cols: &[String],
-) -> Vec<DictionaryEntry> {
-    // Group frequency records by field
-    let mut frequency_by_field: HashMap<String, Vec<&FrequencyRecord>> = HashMap::new();
-    for freq_record in frequency_records {
-        frequency_by_field
-            .entry(freq_record.field.clone())
-            .or_default()
-            .push(freq_record);
-    }
-
-    let mut dictionary_entries = Vec::new();
-
-    for stats_record in stats_records {
-        let field_name = &stats_record.field;
-        let field_frequencies = frequency_by_field
-            .get(field_name)
-            .cloned()
-            .unwrap_or_default();
-
-        // Generate enumeration
-        let enumeration = if stats_record.cardinality <= enum_threshold as u64 {
-            // Check if there's a rank=0 entry (Other category) or <ALL_UNIQUE> value
-            let has_other = field_frequencies
-                .iter()
-                .any(|f| f.rank == 0.0 && !f.value.contains("<ALL_UNIQUE>"));
-            if has_other {
-                String::new()
-            } else {
-                // Enumerate all values (excluding <ALL_UNIQUE>), each on its own line
-                let mut enum_values: Vec<String> = field_frequencies
-                    .iter()
-                    .filter(|f| !f.value.contains("<ALL_UNIQUE>"))
-                    .map(|f| f.value.clone())
-                    .collect();
-                enum_values.sort(); // Sort alphabetically for consistency
-                enum_values.join("\n")
-            }
-        } else {
-            String::new()
-        };
-
-        // Generate examples
-        let examples = if field_frequencies
-            .iter()
-            .any(|f| (f.percentage - 100.0).abs() < 0.0001)
-        {
-            "<ALL_UNIQUE>".to_string()
-        } else {
-            // Get top N values sorted by count descending
-            let mut sorted_freqs = field_frequencies.clone();
-            sorted_freqs.sort_by(|a, b| b.count.cmp(&a.count).then_with(|| a.value.cmp(&b.value)));
-
-            let top_n: Vec<String> = sorted_freqs
-                .iter()
-                .take(num_examples as usize)
-                .map(|f| {
-                    // For frequency bucket entries (rank == 0.0), strip the redundant
-                    // "(n)" count and append "…" to disambiguate from literal values
-                    // with the same name (e.g. bucket "Other… [4,091]" vs literal
-                    // "Other [2,006]")
-                    let raw_value = if f.rank == 0.0 {
-                        let base = if let Some(pos) = f.value.rfind(" (") {
-                            &f.value[..pos]
-                        } else {
-                            &f.value
-                        };
-                        format!("{base}…")
-                    } else {
-                        f.value.clone()
-                    };
-
-                    let v = if truncate_str > 0 && raw_value.chars().count() > truncate_str {
-                        let mut s = raw_value.chars().take(truncate_str).collect::<String>();
-                        s.push('…');
-                        s
-                    } else {
-                        raw_value
-                    };
-                    format!("{} [{}]", v, f.count)
-                })
-                .collect();
-
-            top_n.join("\n")
-        };
-
-        // Collect additional columns for this entry, preserving order
-        let mut entry_addl_cols = IndexMap::new();
-        for col_name in addl_cols {
-            if let Some(value) = stats_record.addl_cols.get(col_name) {
-                entry_addl_cols.insert(col_name.clone(), value.clone());
-            }
-        }
-
-        dictionary_entries.push(DictionaryEntry {
-            name: stats_record.field.clone(),
-            r#type: stats_record.r#type.clone(),
-            label: String::new(),       // Will be filled by LLM
-            description: String::new(), // Will be filled by LLM
-            min: stats_record.min.clone(),
-            max: stats_record.max.clone(),
-            cardinality: stats_record.cardinality,
-            enumeration,
-            null_count: stats_record.nullcount,
-            addl_cols: entry_addl_cols,
-            examples,
-        });
-    }
-
-    dictionary_entries
-}
-
 /// Extract a single JSON value from an LLM response.
 ///
 /// Strategy (in order):
@@ -1542,55 +1182,6 @@ fn extract_json_from_output(output: &str) -> CliResult<serde_json::Value> {
         "Failed to extract JSON content from LLM response. Output: {}",
         if output.is_empty() { "<empty>" } else { output }
     )
-}
-
-/// Parse LLM JSON response to extract Label and Description for each field
-fn parse_llm_dictionary_response(
-    llm_response: &str,
-    field_names: &[String],
-) -> CliResult<HashMap<String, (String, String)>> {
-    let json_value = extract_json_from_output(llm_response)?;
-
-    let mut result = HashMap::new();
-
-    // Parse JSON object
-    if let Some(obj) = json_value.as_object() {
-        for field_name in field_names {
-            if let Some(field_obj) = obj.get(field_name)
-                && let Some(field_map) = field_obj.as_object()
-            {
-                let label = field_map
-                    .get("label")
-                    .and_then(|v| v.as_str())
-                    .unwrap_or("")
-                    .to_string();
-
-                let description = field_map
-                    .get("description")
-                    .and_then(|v| v.as_str())
-                    .unwrap_or("")
-                    .to_string();
-
-                result.insert(field_name.clone(), (label, description));
-            }
-        }
-    }
-
-    Ok(result)
-}
-
-/// Combine code-generated dictionary entries with LLM-generated Label/Description
-fn combine_dictionary_entries(
-    mut code_entries: Vec<DictionaryEntry>,
-    llm_labels_descriptions: &HashMap<String, (String, String)>,
-) -> Vec<DictionaryEntry> {
-    for entry in &mut code_entries {
-        if let Some((label, description)) = llm_labels_descriptions.get(&entry.name) {
-            entry.label = label.clone();
-            entry.description = description.clone();
-        }
-    }
-    code_entries
 }
 
 /// Replace {GENERATED_BY_SIGNATURE} placeholder with actual attribution
